#1 use system properties to configure externally manged proxy
Quarkus' test extension creates a second instance of the test class which is injected with all required references. JUnit5 is not aware of that second instance an passes the one created by itself to any annotated extension, namely the mountebank extension. Hence setting fields on one instance leafs those fields untouched in the other.